+++ opentelemetry-sdk/tests/logs/test_handler.py
@@ -14,6 +14,8 @@
 
 import logging
 import os
+import sys
+import traceback
 import unittest
 from unittest.mock import Mock, patch
 
@@ -47,6 +49,39 @@ class TestLoggingHandler(unittest.TestCase):
         with self.assertLogs(level=logging.WARNING):
             logger.warning("Warning message")
         self.assertEqual(processor.emit_count(), 1)
+
+    def test_handler_error_exc_info(self):
+        processor, logger = set_up_test_logging(logging.NOTSET)
+
+        class CustomException(Exception):
+            pass
+
+        try:
+            raise CustomException("Custom exception")
+        except CustomException as exception:
+            exc_info = (type(exception), exception, exception.__traceback__)
+        else:
+            exc_info = "Second stringified exception"
+
+        exc_info_values = [
+            # Don't know what caused it in my context, so I'm relying on mocks to replicate the behavior.
+            # First the `record.exc_info` becomes a string somehow, then `sys.exc_info` brings the tuple.
+            "First stringified exception",
+            exc_info,
+        ]
+
+        with patch.object(sys, "exc_info", side_effect=exc_info_values):
+            logger.exception("Exception message")  # Should not raise exception
+
+        assert processor.emit_count() == 1
+
+        attributes = processor.log_data_emitted[0].log_record.attributes._dict
+        assert attributes["exception.type"] == "CustomException"
+        assert attributes["exception.message"] == str(exc_info[1])
+        assert isinstance(exc_info, tuple)
+        assert attributes["exception.stacktrace"] == "".join(
+            traceback.format_exception(*exc_info)
+        )
 
     def test_handler_custom_log_level(self):
         processor, logger = set_up_test_logging(logging.ERROR)